# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/user/changepwd": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "修改密码",
                "parameters": [
                    {
                        "description": "修改密码",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UserChangePwdReq"
                        }
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "400": {
                        "description": "错误的请求",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "500": {
                        "description": "系统异常",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    }
                }
            }
        },
        "/api/user/detail/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取用户详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取用户详情",
                        "schema": {
                            "$ref": "#/definitions/responses.UserDetailModel"
                        }
                    },
                    "400": {
                        "description": "错误的请求",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "500": {
                        "description": "系统异常",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    }
                }
            }
        },
        "/api/user/glogin": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "登录",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UserLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/responses.UserLoginRes"
                        }
                    },
                    "400": {
                        "description": "错误的请求",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "500": {
                        "description": "系统异常",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    }
                }
            }
        },
        "/api/user/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取用户列表",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "role",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取用户列表",
                        "schema": {
                            "$ref": "#/definitions/responses.PagedUserList"
                        }
                    },
                    "400": {
                        "description": "错误的请求",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "500": {
                        "description": "系统异常",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "登录",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UserLoginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/responses.UserLoginRes"
                        }
                    },
                    "400": {
                        "description": "错误的请求",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "500": {
                        "description": "系统异常",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    }
                }
            }
        },
        "/api/user/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "description": "注册",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UserRegisterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/responses.IdResult"
                        }
                    },
                    "400": {
                        "description": "错误的请求",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "500": {
                        "description": "系统异常",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    }
                }
            }
        },
        "/api/user/resetpwd": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "重置密码",
                "parameters": [
                    {
                        "description": "重置密码",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UserResetPwdReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "400": {
                        "description": "错误的请求",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "500": {
                        "description": "系统异常",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    }
                }
            }
        },
        "/api/user/sendverifyemail": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "发送认证邮件",
                "parameters": [
                    {
                        "description": "发送认证邮件",
                        "name": "\"\"",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.SendVerifyEmailReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "400": {
                        "description": "错误的请求",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "403": {
                        "description": "无权限",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    },
                    "500": {
                        "description": "系统异常",
                        "schema": {
                            "$ref": "#/definitions/responses.Result"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "requests.SendVerifyEmailReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "requests.UserChangePwdReq": {
            "type": "object",
            "properties": {
                "newpwd": {
                    "type": "string"
                },
                "oldpwd": {
                    "type": "string"
                }
            }
        },
        "requests.UserLoginReq": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "pwd": {
                    "type": "string"
                }
            }
        },
        "requests.UserRegisterReq": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "pwd": {
                    "type": "string"
                }
            }
        },
        "requests.UserResetPwdReq": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
        "responses.IdResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "responses.PagedUserList": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.UserDetailModel"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "responses.Result": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "responses.UserDetailModel": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "integer"
                }
            }
        },
        "responses.UserLoginModel": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "responses.UserLoginRes": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/responses.UserLoginModel"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}